♻️ refactor: graph 함수들 custom hook으로 분리
