+++ frontend/src/hooks/useGraphData.ts
