+++ frontend/src/hooks/useNodeUpdate.ts
@@ -1,0 +1,34 @@
+import * as d3 from 'd3';
+import { useCallback } from 'react';
+
+export default function useNodeUpdate(nodes: any[], changeHoveredNode: (key: string) => void) {
+  return useCallback(
+    (nodesSelector: SVGGElement) => {
+      const normalSymbol = d3.symbol().type(d3.symbolSquare).size(10)();
+      const starSymbol = d3.symbol().type(d3.symbolStar).size(70)();
+
+      d3.select(nodesSelector)
+        .selectAll('path')
+        .data(nodes)
+        .join('path')
+        .attr('transform', (d) => `translate(${[d.x, d.y]})`)
+        .attr('d', (d) => (d.isSelected ? starSymbol : normalSymbol));
+
+      d3.select(nodesSelector)
+        .selectAll('text')
+        .data(nodes)
+        .join('text')
+        .text((d) => d.author)
+        .on('mouseover', (e, d) => {
+          changeHoveredNode(d.key);
+        })
+        .on('mouseout', () => {
+          changeHoveredNode('');
+        })
+        .attr('x', (d) => d.x)
+        .attr('y', (d) => d.y + 10)
+        .attr('dy', 5);
+    },
+    [nodes, changeHoveredNode],
+  );
+}
